Refactor structure

+++ src/com/ubhave/datahandler/transfer/DataTransfer.java
@@ -13,6 +13,7 @@ import android.preference.PreferenceManager;
 import android.util.Log;
 
 import com.ubhave.datahandler.config.DataHandlerConfig;
+import com.ubhave.datahandler.config.DataStorageConstants;
 import com.ubhave.datahandler.config.DataTransferConfig;
 import com.ubhave.datahandler.except.DataHandlerException;
 import com.ubhave.datahandler.http.WebConnection;
@@ -21,7 +22,7 @@ public class DataTransfer implements DataTransferInterface
 {
 	private final static String TAG = "DataTransfer";
 	private static final String LAST_LOGS_UPLOAD_TIME = "com.ubhave.datahandler.LAST_LOGS_UPLOAD_TIME";
-	
+
 	private final Context context;
 	private final DataHandlerConfig config;
 
@@ -31,43 +32,62 @@ public class DataTransfer implements DataTransferInterface
 		this.config = DataHandlerConfig.getInstance();
 		setLogsUploadTime(System.currentTimeMillis());
 	}
-	
+
 	@Override
 	public void uploadData(final String uploadDirectory) throws DataHandlerException
 	{
+		if (DataHandlerConfig.shouldLog())
+		{
+			Log.d(TAG, "Attempting upload from: " + uploadDirectory);
+		}
 		File directory = new File(uploadDirectory);
 		File[] files = directory.listFiles();
 		if (files != null)
 		{
+			if (DataHandlerConfig.shouldLog())
+			{
+				Log.d(TAG, "Attempting upload "+files.length+" files.");
+			}
 			for (File file : files)
 			{
-				HashMap<String, String> paramsMap = getPostParams();
-				String url = (String) config.get(DataTransferConfig.POST_DATA_URL);
-				if (DataHandlerConfig.shouldLog())
+				if (file.isFile() && file.getName().contains(DataStorageConstants.ZIP_FILE_SUFFIX))
 				{
-					Log.d(TAG, "Posting to: "+url);
-				}
-				
-				String response = WebConnection.postDataToServer(url, file, paramsMap);
-				if (response.equals(config.get(DataTransferConfig.POST_RESPONSE_ON_SUCCESS)))
-				{
+					HashMap<String, String> paramsMap = getPostParams();
+					String url = (String) config.get(DataTransferConfig.POST_DATA_URL);
 					if (DataHandlerConfig.shouldLog())
 					{
-						Log.d(TAG, "file " + file + " successfully uploaded to the server");
-						Log.d(TAG, "file " + file + " deleting local copy");
+						Log.d(TAG, "Posting to: " + url);
 					}
-					file.delete();
-					setLogsUploadTime(System.currentTimeMillis());
-				}
-				else 
-				{
-					if (DataHandlerConfig.shouldLog())
+
+					String response = WebConnection.postDataToServer(url, file, paramsMap);
+					if (response.equals(config.get(DataTransferConfig.POST_RESPONSE_ON_SUCCESS)))
 					{
-						Log.d(TAG, "file " + file + " failed to upload file to the server, response received: " + response);
+						if (DataHandlerConfig.shouldLog())
+						{
+							Log.d(TAG, "file " + file + " successfully uploaded to the server");
+							Log.d(TAG, "file " + file + " deleting local copy");
+						}
+						file.delete();
+						setLogsUploadTime(System.currentTimeMillis());
 					}
-					throw new DataHandlerException(DataHandlerException.POST_FAILED);
+					else
+					{
+						if (DataHandlerConfig.shouldLog())
+						{
+							Log.d(TAG, "file " + file + " failed to upload file to the server, response received: " + response);
+						}
+						throw new DataHandlerException(DataHandlerException.POST_FAILED);
+					}
+				}
+				else if (DataHandlerConfig.shouldLog())
+				{
+					Log.d(TAG, "Skip: " + file.getName());
 				}
 			}
+		}
+		else if (DataHandlerConfig.shouldLog())
+		{
+			Log.d(TAG, "Attempting file list is null.");
 		}
 	}
 
@@ -97,7 +117,7 @@ public class DataTransfer implements DataTransferInterface
 		SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context);
 		return preferences.getLong(LAST_LOGS_UPLOAD_TIME, 0);
 	}
-	
+
 	private HashMap<String, String> getPostParams()
 	{
 		HashMap<String, String> paramsMap = new HashMap<String, String>();
@@ -128,7 +148,7 @@ public class DataTransfer implements DataTransferInterface
 		}
 		return paramsMap;
 	}
-	
+
 	private void post(final String url, final String data) throws DataHandlerException
 	{
 		try
@@ -136,12 +156,12 @@ public class DataTransfer implements DataTransferInterface
 			String dataKey = (String) config.get(DataTransferConfig.POST_RAW_DATA_KEY);
 			JSONObject dataParam = new JSONObject();
 			dataParam.put(dataKey, data);
-			
+
 			HashMap<String, String> paramsMap = getPostParams();
 			paramsMap.put(dataKey, data);
-			
+
 			String response = WebConnection.postToServer(url, paramsMap);
-			String expectedResponse = (String) config.get(DataTransferConfig.POST_RESPONSE_ON_SUCCESS); 
+			String expectedResponse = (String) config.get(DataTransferConfig.POST_RESPONSE_ON_SUCCESS);
 			if (!response.equals(expectedResponse))
 			{
 				throw new DataHandlerException(DataHandlerException.POST_FAILED);
